+++ mobile/app/src/main/java/com/bellako/kiwi/features/goals/data/GoalState.kt
@@ -2,7 +2,7 @@ package com.bellako.kiwi.features.goals.data
 
 data class GoalState(
     val id: String = "",
-    val objective: Long = 0,
+    val objective: Int = 0,
     val description: String = "",
     val type: String = "EXERCISE",
     val category: String = "DAILY_CHALLENGES",
